Expose moc-building methods, and add documentation notebook.
